#assign1 -changed float to double #assign2 -added bst and some of its functions

+++ Assign_1/Assign_1/main.c
@@ -85,7 +85,7 @@ void *scanInteger(FILE *file) {
 }
 
 void *scanReal(FILE *file) {
-	float token = readReal(file);
+	double token = readReal(file);
 	if (token == 0) {
 		return 0;
 	}
